Restaurant detail view added

+++ app/src/main/java/com/example/restaurants/models/Review.kt
@@ -1,3 +1,3 @@
 package com.example.restaurants.models
 
-data class Review (val id: String, val userName: String, val text: String, val rating: Double)
+data class Review (val id: String, val userName: String, val text: String, val rating: Float)
